+++ releasenotes.md
@@ -98,6 +98,7 @@
 - Memory leak in Object when not using temp allocators.
 - Tracking allocator would double the allocations in the report.
 - `printf` will now show errors in the output when there are errors.
+- Bug where `if try` would work incorrectly in a macro.
 
 ### Stdlib changes